\name{vaso}
\alias{vaso}
\docType{data}
\title{Vaso data}
\usage{data(vaso)}
\description{Finney's data on vaso-contriction in the skin of the digits. \cr
The \code{vaso} data frame has 39 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{volume}{volume}
    \item{rate}{rate}
    \item{y}{response: 0= nonoccurrence, 1= occurrence}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.23}
%\examples{}
\keyword{datasets}
